Refactor code into classes
