+++ tools/build.gradle.kts
@@ -58,7 +58,7 @@ tasks.withType<JavaExec> {
 compose {
     desktop {
         application {
-            mainClass = "world.gregs.voidps.tools.MainKt" // The entry point of your app
+            mainClass = "world.gregs.voidps.tools.search.MainKt" // The entry point of your app
             nativeDistributions {
                 targetFormats(TargetFormat.Msi, TargetFormat.Deb)
                 packageName = "MyComposeApp"
