+++ tools/src/main/kotlin/world/gregs/voidps/tools/search/Theme.kt
@@ -1,0 +1,22 @@
+package world.gregs.voidps.tools.search
+
+import androidx.compose.ui.graphics.Color
+
+internal val BgDark = Color(0xFF0F1117)
+internal val BgPanel = Color(0xFF161B27)
+internal val BgCard = Color(0xFF1E2535)
+internal val BgSelected = Color(0xFF1A3A5C)
+internal val BgMultiSelected = Color(0xFF122840)
+internal val AccentBlue = Color(0xFF4A9EFF)
+internal val AccentLight = Color(0xFF7BBEFF)
+internal val TextPrimary = Color(0xFFE8EDF5)
+internal val TextSecond = Color(0xFF8A95A8)
+internal val TextMuted = Color(0xFF4A5568)
+internal val BorderColor = Color(0xFF2A3347)
+internal val TagBg = Color(0xFF1A2840)
+internal val TagText = Color(0xFF5BA3E8)
+internal val SuccessGreen = Color(0xFF3DD68C)
+internal val WarnAmber = Color(0xFFFFB547)
+internal val ParamKey = Color(0xFFBF94E4)
+internal val ParamVal = Color(0xFFE5C07B)
+internal val LinkColor = Color(0xFF56B6C2)
