+++ tools/src/main/kotlin/world/gregs/voidps/tools/search/screen/AppScreen.kt
@@ -1,0 +1,6 @@
+package world.gregs.voidps.tools.search.screen
+
+enum class AppScreen {
+    PICKER,
+    BROWSER
+}
