+++ tools/src/main/kotlin/world/gregs/voidps/tools/search/screen/view/detail/FieldLink.kt
@@ -1,0 +1,6 @@
+package world.gregs.voidps.tools.search.screen.view.detail
+
+data class FieldLink(
+    val fieldName: String,       // field on source, e.g. "runSound"
+    val targetTabLabel: String,  // label of destination tab, e.g. "Sounds"
+)
