+++ tools/src/main/kotlin/world/gregs/voidps/tools/search/screen/view/table/filter/FieldFilter.kt
@@ -1,0 +1,7 @@
+package world.gregs.voidps.tools.search.screen.view.table.filter
+
+data class FieldFilter(
+    val fieldName: String,
+    val value: String = "",
+    val mode: MatchMode = MatchMode.CONTAINS,
+)
