+++ tools/src/main/kotlin/world/gregs/voidps/tools/search/screen/view/table/filter/MatchMode.kt
@@ -1,0 +1,13 @@
+package world.gregs.voidps.tools.search.screen.view.table.filter
+
+enum class MatchMode {
+    CONTAINS,
+    EXACT,
+    GREATER_THAN,
+    LESS_THAN,
+    HAS_VALUE,
+    PARAM_KEY,
+    PARAM_VALUE,
+    NOT_NULL,
+    NOT_EMPTY
+}
