+++ tools/src/main/kotlin/world/gregs/voidps/tools/search/screen/view/table/filter/SearchField.kt
@@ -1,0 +1,42 @@
+package world.gregs.voidps.tools.search.screen.view.table.filter
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.BasicTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.SolidColor
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import world.gregs.voidps.tools.search.AccentBlue
+import world.gregs.voidps.tools.search.BgCard
+import world.gregs.voidps.tools.search.BorderColor
+import world.gregs.voidps.tools.search.TextMuted
+import world.gregs.voidps.tools.search.TextPrimary
+
+@Composable
+fun SearchField(value: String, onValueChange: (String) -> Unit, placeholder: String, modifier: Modifier = Modifier) {
+    Box(
+        modifier = modifier
+            .height(28.dp)
+            .background(BgCard, RoundedCornerShape(4.dp))
+            .border(0.5.dp, BorderColor, RoundedCornerShape(4.dp))
+            .padding(horizontal = 8.dp),
+        contentAlignment = Alignment.CenterStart,
+    ) {
+        if (value.isEmpty()) Text(placeholder, color = TextMuted, fontSize = 12.sp)
+        BasicTextField(
+            value, onValueChange, singleLine = true,
+            textStyle = TextStyle(color = TextPrimary, fontSize = 12.sp),
+            cursorBrush = SolidColor(AccentBlue), modifier = Modifier.fillMaxWidth()
+        )
+    }
+}
